Fixed warnings in Raspberry Pi

+++ example/example_oui.c
@@ -1237,7 +1237,7 @@ int main()
     
     init(_vg);
 
-    printf("sizeof(UIitem)=%lu\n", sizeof(UIitem));
+    printf("sizeof(UIitem)=%lu\n", (long unsigned int)sizeof(UIitem));
 
     glfwSwapInterval(0);
 
@@ -1288,7 +1288,7 @@ int main()
         glfwPollEvents();
     }
     printf("Peak item count: %i (%lu bytes)\nPeak allocated handles: %u bytes\n",
-            peak_items, peak_items * sizeof(UIitem), peak_alloc);
+            peak_items, peak_items * (long unsigned int)sizeof(UIitem), peak_alloc);
 
     uiDestroyContext(uictx);
